feat: created global websocket subscription to ensure all notifications are handled in one place, reduce API calling and polling, automatically reconnect websocket, and improve synchronization.
